# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrainHub360 corporate training management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: workspace
    description: Current organization and workspace context
  - name: dashboard
    description: Programme and organization KPI summaries
  - name: learning
    description: Courses, programmes, and competency assessments
  - name: engagement
    description: Messaging, bookings, and calendar
  - name: auth
  - name: enrolments
  - name: analytics
  - name: governance
  - name: review-credit
    description: Optional performance review credit module
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /workspace:
    get:
      operationId: getWorkspace
      tags: [workspace]
      summary: Get current workspace context
      responses:
        "200":
          description: Current workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get KPI dashboard summary
      responses:
        "200":
          description: Dashboard KPI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /programmes:
    get:
      operationId: listProgrammes
      tags: [learning]
      summary: List programmes visible to the current user
      responses:
        "200":
          description: Programmes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Programme"
  /courses:
    get:
      operationId: listCourses
      tags: [learning]
      summary: List courses in the active workspace
      parameters:
        - name: programmeId
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [learning]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}/attempts:
    post:
      operationId: submitAssessmentAttempt
      tags: [learning]
      summary: Submit an assessment attempt
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentAttemptInput"
      responses:
        "200":
          description: Assessment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
  /sessions:
    get:
      operationId: listSessions
      tags: [engagement]
      summary: List upcoming training sessions
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingSession"
  /threads:
    get:
      operationId: listMessageThreads
      tags: [engagement]
      summary: List message threads for the active workspace
      responses:
        "200":
          description: Message threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageThread"
  /threads/{threadId}/messages:
    get:
      operationId: listThreadMessages
      tags: [engagement]
      summary: List messages in a thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreadMessage"
    post:
      operationId: sendThreadMessage
      tags: [engagement]
      summary: Send a message in a thread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreadMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadMessage"
  /bookings:
    get:
      operationId: listBookings
      tags: [engagement]
      summary: List one-to-one bookings
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [engagement]
      summary: Book a one-to-one session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /calendar:
    get:
      operationId: listCalendarEvents
      tags: [engagement]
      summary: List the unified training calendar
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
  /review-credit/settings:
    get:
      operationId: getReviewCreditSettings
      tags: [review-credit]
      summary: Get review credit configuration
      responses:
        "200":
          description: Review credit settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCreditSettings"
    patch:
      operationId: updateReviewCreditSettings
      tags: [review-credit]
      summary: Update review credit configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreditSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewCreditSettings"
  /review-credit/export:
    get:
      operationId: exportReviewCredits
      tags: [review-credit]
      summary: Export review credit evidence as CSV
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email/password (JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: JWT tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/sso:
    post:
      operationId: ssoLogin
      tags: [auth]
      summary: SSO/SAML token exchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SsoRequest"
      responses:
        "200":
          description: JWT tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /enrolments:
    get:
      operationId: listEnrolments
      tags: [enrolments]
      summary: List enrolments
      responses:
        "200":
          description: Enrolments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrolment"
    post:
      operationId: createEnrolment
      tags: [enrolments]
      summary: Create enrolment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrolmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
  /enrolments/{enrolmentId}:
    patch:
      operationId: updateEnrolment
      tags: [enrolments]
      summary: Update enrolment (recomputes score, audit-logged)
      parameters:
        - name: enrolmentId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrolmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrolment"
    delete:
      operationId: deleteEnrolment
      tags: [enrolments]
      summary: Delete enrolment (requires second-admin approval per governance)
      parameters:
        - name: enrolmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /enrolments/{enrolmentId}/score:
    get:
      operationId: getEnrolmentScore
      tags: [enrolments]
      summary: Get computed review-credit score for an enrolment
      parameters:
        - name: enrolmentId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Score breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolmentScore"
  /programmes/{programmeId}/modules:
    get:
      operationId: listModules
      tags: [learning]
      summary: List modules in a programme course set
      parameters:
        - name: programmeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [learning]
      summary: Create module (micro-learning unit)
      parameters:
        - name: programmeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{moduleId}/assessments:
    get:
      operationId: listAssessments
      tags: [learning]
      summary: List assessments (recall vs applied)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [learning]
      summary: Create assessment tagged recall|applied
      parameters:
        - name: moduleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /kpi/summary:
    get:
      operationId: getKpiSummary
      tags: [analytics]
      summary: KPI aggregates (competencyRate, dropOff, timeToCompetency, trainerUtilization)
      responses:
        "200":
          description: KPI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /audit/logs:
    get:
      operationId: listAuditLogs
      tags: [governance]
      summary: Query audit trail (admin only)
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
  /nudge/run:
    post:
      operationId: runNudge
      tags: [engagement]
      summary: Trigger calendar-gap nudge pass (internal)
      responses:
        "200":
          description: Nudges sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NudgeResult"
  /help/ai:
    post:
      operationId: aiHelp
      tags: [engagement]
      summary: In-context AI help stub for stuck learners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequest"
      responses:
        "200":
          description: Help answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpResponse"
  /organizations/{orgId}/review-credits:
    get:
      operationId: getReviewCredits
      tags: [review-credit]
      summary: REST export of review-credit aggregates for HRIS
      parameters:
        - name: orgId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Review credit rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewCreditRow"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Workspace:
      type: object
      required: [organization, activeProgramme, user, dataSource]
      properties:
        organization:
          $ref: "#/components/schemas/Organization"
        activeProgramme:
          $ref: "#/components/schemas/Programme"
        user:
          $ref: "#/components/schemas/CurrentUser"
        dataSource:
          type: string
    Organization:
      type: object
      required: [id, name, plan, learnerCount, programmeCount]
      properties:
        id: { type: string }
        name: { type: string }
        plan: { type: string }
        learnerCount: { type: integer }
        programmeCount: { type: integer }
    CurrentUser:
      type: object
      required: [id, name, email, role, initials]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        initials: { type: string }
    Programme:
      type: object
      required: [id, name, type, status, learnerCount, courseCount, progress, owner]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        learnerCount: { type: integer }
        courseCount: { type: integer }
        progress: { type: number }
        owner: { type: string }
    Course:
      type: object
      required: [id, title, programmeId, programmeName, category, progress, status, appliedThreshold, appliedScore, dueDate, trainer]
      properties:
        id: { type: string }
        title: { type: string }
        programmeId: { type: string }
        programmeName: { type: string }
        category: { type: string }
        progress: { type: number }
        status: { type: string }
        appliedThreshold: { type: number }
        appliedScore: { type: ["number", "null"] }
        dueDate: { type: ["string", "null"] }
        trainer: { type: string }
        duration: { type: ["string", "null"] }
    CourseInput:
      type: object
      required: [title, programmeId, category]
      properties:
        title: { type: string }
        programmeId: { type: string }
        category: { type: string }
        appliedThreshold: { type: number }
        dueDate: { type: ["string", "null"] }
    DashboardSummary:
      type: object
      required: [complianceHealth, completionRate, competencyRate, expiringCredentials, dropOffRate, timeToCompetency, trainerUtilization, weeklyActivity, dropOffHeatmap, expiringItems]
      properties:
        complianceHealth: { type: number }
        completionRate: { type: number }
        competencyRate: { type: number }
        expiringCredentials: { type: integer }
        dropOffRate: { type: number }
        timeToCompetency: { type: number }
        trainerUtilization: { type: number }
        weeklyActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityPoint"
        dropOffHeatmap:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapPoint"
        expiringItems:
          type: array
          items:
            $ref: "#/components/schemas/ExpiringCredential"
    ActivityPoint:
      type: object
      required: [label, active, completed]
      properties:
        label: { type: string }
        active: { type: integer }
        completed: { type: integer }
    HeatmapPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    ExpiringCredential:
      type: object
      required: [name, course, expires, status]
      properties:
        name: { type: string }
        course: { type: string }
        expires: { type: string }
        status: { type: string }
    TrainingSession:
      type: object
      required: [id, title, programme, date, time, mode, attendees, status]
      properties:
        id: { type: string }
        title: { type: string }
        programme: { type: string }
        date: { type: string }
        time: { type: string }
        mode: { type: string }
        location: { type: ["string", "null"] }
        videoLink: { type: ["string", "null"] }
        attendees: { type: integer }
        status: { type: string }
    MessageThread:
      type: object
      required: [id, subject, context, participant, preview, updatedAt, unread, urgent]
      properties:
        id: { type: string }
        subject: { type: string }
        context: { type: string }
        participant: { type: string }
        preview: { type: string }
        updatedAt: { type: string }
        unread: { type: boolean }
        urgent: { type: boolean }
    ThreadMessage:
      type: object
      required: [id, threadId, author, body, createdAt, urgent]
      properties:
        id: { type: string }
        threadId: { type: string }
        author: { type: string }
        body: { type: string }
        createdAt: { type: string }
        urgent: { type: boolean }
    ThreadMessageInput:
      type: object
      required: [body]
      properties:
        body: { type: string }
        urgent: { type: boolean }
    Booking:
      type: object
      required: [id, trainer, course, date, time, duration, status]
      properties:
        id: { type: string }
        trainer: { type: string }
        course: { type: string }
        date: { type: string }
        time: { type: string }
        duration: { type: string }
        status: { type: string }
        meetingLink: { type: ["string", "null"] }
    BookingInput:
      type: object
      required: [trainer, date, time]
      properties:
        trainer: { type: string }
        course: { type: string }
        date: { type: string }
        time: { type: string }
    CalendarEvent:
      type: object
      required: [id, title, date, time, type, accent]
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string }
        time: { type: string }
        type: { type: string }
        accent: { type: string }
        location: { type: ["string", "null"] }
        meetingLink: { type: ["string", "null"] }
    AssessmentAttemptInput:
      type: object
      required: [appliedScore, recallScore]
      properties:
        appliedScore: { type: number, minimum: 0, maximum: 100 }
        recallScore: { type: number, minimum: 0, maximum: 100 }
    AssessmentResult:
      type: object
      required: [appliedScore, recallScore, passed, competenceMet, message]
      properties:
        appliedScore: { type: number }
        recallScore: { type: number }
        passed: { type: boolean }
        competenceMet: { type: boolean }
        message: { type: string }
    ReviewCreditSettings:
      type: object
      required: [enabled, maxWeighting, assessmentWeight, timelinessWeight, applicationWeight, requireManagerSignoff, collectApplicationScores, eligibleProgrammeTypes]
      properties:
        enabled: { type: boolean }
        maxWeighting: { type: number }
        assessmentWeight: { type: number }
        timelinessWeight: { type: number }
        applicationWeight: { type: number }
        requireManagerSignoff: { type: boolean }
        collectApplicationScores: { type: boolean }
        eligibleProgrammeTypes:
          type: array
          items: { type: string }
    ReviewCreditSettingsUpdate:
      type: object
      properties:
        enabled: { type: boolean }
        maxWeighting: { type: number }
        assessmentWeight: { type: number }
        timelinessWeight: { type: number }
        applicationWeight: { type: number }
        collectApplicationScores: { type: boolean }
        eligibleProgrammeTypes:
          type: array
          items: { type: string }

    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken: { type: string }
        user: { $ref: "#/components/schemas/CurrentUser" }
    SsoRequest:
      type: object
      required: [samlAssertion]
      properties:
        samlAssertion: { type: string }
    Enrolment:
      type: object
      required: [id, learnerId, courseId, status]
      properties:
        id: { type: string }
        learnerId: { type: string }
        courseId: { type: string }
        status: { type: string }
        enrolmentScore: { type: ["number", "null"] }
        managerSignOff: { type: ["boolean", "null"] }
    EnrolmentInput:
      type: object
      required: [learnerId, courseId]
      properties:
        learnerId: { type: string }
        courseId: { type: string }
        deadlineDate: { type: ["string", "null"] }
    EnrolmentUpdate:
      type: object
      properties:
        status: { type: string }
        appliedAssessmentScore: { type: number }
        applicationScore: { type: number }
        managerSignOff: { type: boolean }
    EnrolmentScore:
      type: object
      required: [enrolmentId, score, breakdown]
      properties:
        enrolmentId: { type: string }
        score: { type: number }
        breakdown: { type: object }
    Module:
      type: object
      required: [id, courseId, title]
      properties:
        id: { type: string }
        courseId: { type: string }
        title: { type: string }
    ModuleInput:
      type: object
      required: [courseId, title]
      properties:
        courseId: { type: string }
        title: { type: string }
    Assessment:
      type: object
      required: [id, moduleId, type]
      properties:
        id: { type: string }
        moduleId: { type: string }
        type: { type: string, enum: [recall, applied] }
    AssessmentInput:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [recall, applied] }
        maxScore: { type: number }
    AuditLog:
      type: object
      required: [id, action, entity, timestamp]
      properties:
        id: { type: string }
        action: { type: string }
        entity: { type: string }
        timestamp: { type: string }
    NudgeResult:
      type: object
      required: [sent]
      properties:
        sent: { type: integer }
    HelpRequest:
      type: object
      required: [question]
      properties:
        question: { type: string }
        courseId: { type: string }
    HelpResponse:
      type: object
      required: [answer]
      properties:
        answer: { type: string }
    ReviewCreditRow:
      type: object
      required: [learnerId, raw, final]
      properties:
        learnerId: { type: string }
        raw: { type: number }
        final: { type: number }